Made changes to fix performance issue.
Root Cause :- It is performing poor because we were using string to add the item from list and
display on HTML. String always create new instance when we add new item to string which eventually decrease the performance.

Stringbuilder :- Doesnt create new instance when we add new item to string builder.

String was good option if we have only one item to add but here we are getting many items from list and adding one by one which eventually creating instance everytime and performing poor.